LibMaps: Move MapWidget to new LibMaps library
