+++ Userland/Applications/Maps/UsersMapWidget.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "MapWidget.h"
+#include <LibMaps/MapWidget.h>
 
 namespace Maps {
