+++ Userland/Libraries/CMakeLists.txt
@@ -41,6 +41,7 @@ add_subdirectory(LibLine)
 add_subdirectory(LibLocale)
 add_subdirectory(LibMain)
 add_subdirectory(LibManual)
+add_subdirectory(LibMaps)
 add_subdirectory(LibMarkdown)
 add_subdirectory(LibPartition)
 add_subdirectory(LibPCIDB)
